// packages/budget-sync-client/src/splitwise/codegen/types.gen.ts
export type GetGetExpensesData = {
	body?: never;
	path?: never;
	query?: {
		/**
		 * If provided, only expenses in that group will be returned, and `friend_id` will be ignored.
		 */
		group_id?: number;
		/**
		 * ID of another user. If provided, only expenses between the current and provided user will be returned.
		 */
		friend_id?: number;
		dated_after?: string;
		dated_before?: string;
		updated_after?: string;
		updated_before?: string;
		limit?: number;
		offset?: number;
	};
	url: "/get_expenses";
};